moved function out to module
